requirements.txt
python-dotenv==1.0.1
Requests==2.32.3
twitchAPI==4.3.1
twitchAPI==4.3.1
zebra==0.1.0
zebrafy==1.2.2
